Add SplashMixin
The mixin contains the code that's shared between the Splash screen's
implementations on r2 and Thunderstore mod managers.

Refs TS-1140
